Refactor get relationships endpoint into new RelationshipController
