+++ SocialCareCaseViewerApi/V1/Controllers/SocialCareCaseViewerApiController.cs
@@ -25,12 +25,11 @@ namespace SocialCareCaseViewerApi.V1.Controllers
         private readonly IWarningNoteUseCase _warningNoteUseCase;
         private readonly IGetVisitByVisitIdUseCase _getVisitByVisitIdUseCase;
         private readonly IPersonUseCase _personUseCase;
-        private readonly IRelationshipsV1UseCase _relationshipsV1UseCase;
 
         public SocialCareCaseViewerApiController(IGetAllUseCase getAllUseCase, IAddNewResidentUseCase addNewResidentUseCase,
             IAllocationsUseCase allocationUseCase, ICaseNotesUseCase caseNotesUseCase,
             IVisitsUseCase visitsUseCase, IWarningNoteUseCase warningNotesUseCase,
-            IGetVisitByVisitIdUseCase getVisitByVisitIdUseCase, IPersonUseCase personUseCase, IRelationshipsV1UseCase relationshipsV1UseCase)
+            IGetVisitByVisitIdUseCase getVisitByVisitIdUseCase, IPersonUseCase personUseCase)
         {
             _getAllUseCase = getAllUseCase;
             _addNewResidentUseCase = addNewResidentUseCase;
@@ -40,7 +39,6 @@ namespace SocialCareCaseViewerApi.V1.Controllers
             _warningNoteUseCase = warningNotesUseCase;
             _getVisitByVisitIdUseCase = getVisitByVisitIdUseCase;
             _personUseCase = personUseCase;
-            _relationshipsV1UseCase = relationshipsV1UseCase;
         }
 
         /// <summary>
@@ -378,28 +376,6 @@ namespace SocialCareCaseViewerApi.V1.Controllers
             catch (PatchWarningNoteException e)
             {
                 return BadRequest(e.Message);
-            }
-        }
-
-        /// <summary>
-        /// Get a list of relationships by person id
-        /// </summary>
-        /// <param name="request"></param>
-        /// <response code="200">Successful request. Relationships returned</response>
-        /// <response code="404">Person not found</response>
-        /// <response code="500">There was a problem getting the relationships</response>
-        [ProducesResponseType(typeof(ListRelationshipsV1Response), StatusCodes.Status200OK)]
-        [HttpGet]
-        [Route("residents/{personId}/relationships")]
-        public IActionResult ListRelationships([FromQuery] ListRelationshipsV1Request request)
-        {
-            try
-            {
-                return Ok(_relationshipsV1UseCase.ExecuteGet(request));
-            }
-            catch (GetRelationshipsException ex)
-            {
-                return NotFound(ex.Message);
             }
         }
     }
